Feat(Parser): parse use statements
I'd like to note that this commit launches a dear tradition of failing commits,
where we push *failing* commits into the branch we're working on so as to have
more motivation to fix those failing tests in the future and to know for sure,
upon return, where to start!

For example, in this commit, I've found that even though Parser can easily
parse something like

    use core:io;

it doesn't have an ability to parse

    mod core:io;

which is a bug... not a feature. Thus, I left the tcParseMod test group failing
so as to know exactly what to do when I sit behind the screen next time.