fix(parser): 'use statement after definition' test
However, this fix has demonstrated that the current parser architecture is
deficient in terms of its ability to choose appropriate errors.

In the following commits, we will address that issue by taking advice from the
paper that gave birth to the Parsec library:

    https://www.microsoft.com/en-us/research/wp-content/uploads/2016/02/parsec-paper-letter.pdf

LL(1) grammar has to be implemented using the updated (>>=) and (<|>)
operators to insure efficiency and correctness.